% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakPicking.R
\name{alignment_BfG_cpp}
\alias{alignment_BfG_cpp}
\title{Alignment function}
\usage{
alignment_BfG_cpp(peaklists, mz_dev, DeltaRT, mz_dev_unit)
}
\description{
This function is central to the entire workflow. The main function is programmed in cpp
make sure to also do the summarize_groups function afterwards
}
