% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakPicking.R
\name{alignment_BfG_cpp_par}
\alias{alignment_BfG_cpp_par}
\title{Alignment parallel (wrapper)}
\usage{
alignment_BfG_cpp_par(
  peaklists,
  ppm_dev,
  DeltaRT,
  mz_dev_unit,
  mDa_split = 100,
  numSplits = 16
)
}
\description{
Alignment parallel (wrapper)
}
