% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suspect_screening.R
\name{annotate_grouped}
\alias{annotate_grouped}
\title{Annotate alignment table from non-target app by MS2 searching}
\usage{
annotate_grouped(
  sampleListLocal,
  peakListList,
  alignmentTable,
  db_path,
  threshold_score = 500,
  mztolu = 0.005,
  rttol = 1,
  polarity = "pos",
  CE = c(30, 40),
  CES = c(0, 15),
  instrument = c("LC-ESI-QTOF TripleTOF 5600 SCIEX", "LC-ESI-QTOF TripleTOF 6600 SCIEX",
    "LC-ESI-Orbitrap QExactive"),
  chrom_method = "dx.doi.org/10.1016/j.chroma.2015.11.014",
  ndp_m = 2,
  ndp_n = 1,
  mztolu_ms2 = 0.015,
  rtoffset = 0,
  intCutData = 0,
  numcores = 1,
  datenListLocal,
  expGroups = "BfG"
)
}
\arguments{
\item{peakListList}{list of data.frames of the peak-lists}

\item{alignmentTable}{Knonw as "grouped" in the app}

\item{db_path}{Path to the spectral library (sqlite)}

\item{threshold_score}{dot product threshold score}

\item{mztolu}{m/z tolerance in Da}

\item{rttol}{RT tolerance in min.}

\item{polarity}{Polarity of the measurement}

\item{CE}{Collision energy}

\item{CES}{Collision energy spread}

\item{instrument}{Instruments allowed, must match the instruments listed in the library}

\item{chrom_method}{Chromatography name to choose the correct retention times from the library}

\item{ndp_m}{Peak intensity weighting factor for dot-product}

\item{ndp_n}{m/z weighting factor for dot-product}

\item{mztolu_ms2}{m/z window for dot-product}

\item{rtoffset}{Offset the database to make it match your values (in min)}

\item{intCutData}{Cut off intensity under which fragments are ignored (in data spectra)}

\item{numcores}{Number of cores for parallelization (currently not used)}

\item{sampleList}{data.frame of filenames and paths}
}
\description{
Function takes all the variables needed from non-target app and settings for
searching the database. It returns a table with the annotations for rows of
the alignment table.
}
\details{
datenList is taken from parent frame

If data files are not in memory, they will be temporary loaded and
}
