% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/componentization.R
\name{annotate_grouped_components}
\alias{annotate_grouped_components}
\title{annotate_grouped_components}
\usage{
annotate_grouped_components(
  sampleListLocal,
  peakListList,
  alignmentTable,
  numcores
)
}
\arguments{
\item{sampleListLocal}{(list of processed samples)}

\item{peakListList}{(result of peakpicking)}

\item{alignmentTable}{(result of the alignment)}

\item{numcores}{(number of CPU cores to be used)}
}
\value{
alignmentTable with annotated isotopologues based on componentization
}
\description{
If no db is chosen, annotate alignmentTable with peaklist component info (Cl and Br)
}
