% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library-screening.R
\name{annotate_grouped_mz_rt}
\alias{annotate_grouped_mz_rt}
\title{Annotate alignment table based only on m/z and RT comparison}
\usage{
annotate_grouped_mz_rt(alig, compLibPath, mztol, rttol)
}
\arguments{
\item{alig}{alignment table from shiny app}

\item{compLibPath}{path to csv file}

\item{mztol}{tolerance in u}

\item{rttol}{tolerance in minutes}
}
\value{
Annotation table (`data.frame`) which gives the annotations of the
  alignment table linked by `alignmentID`
}
\description{
Enables also annotation based only on m/z. Uses a csv file substance
library.
}
