% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentTableFilters.R
\name{averageReplicates}
\alias{averageReplicates}
\title{Get mean intensities for replicate injections}
\usage{
averageReplicates(alignment, samples, reps)
}
\arguments{
\item{alignment}{alignment table}

\item{samples}{numeric sample IDs to merge, length must be divisable by repNum}

\item{reps}{number of replicates}
}
\value{
An alignment table with a reduced number of samples but the same number of rows.
}
\description{
This function will take an alignment table and average the intensities of replicate injections.
The first sample in each replicate will remain in the table with average int. Zeros will be
ignored (unless all intensities are zero).
}
