% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment-table-filters.R
\name{blankCorrection}
\alias{blankCorrection}
\title{Remove features also found in the blanks}
\usage{
blankCorrection(
  alignment,
  samplels,
  intensityFactor = 10,
  deleteGrouped = TRUE
)
}
\arguments{
\item{alignment}{Alignment table (matrix)}

\item{samplels}{sample table}

\item{intensityFactor}{Intensity difference between samples and blanks beyond
which features are not deleted}

\item{deleteGrouped}{If true (default) then when a blank feature is found, it
will be removed along with the whole component group (isotopologues,
adducts, in-source fragments)}
}
\value{
An alignment table with rows removed
}
\description{
Which samples are blanks is given in the sample table. The
  features found both in samples and blanks are removed from the alignment
  table, unless the intensity in the samples is higher than the blank by the
  intensityFactor.
}
\details{
If more than one blank sample exists then the highest intensity will
be used for the comparison.
}
