% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{blankCorrection_variable_headerList}
\alias{blankCorrection_variable_headerList}
\title{blankCorrection_variable_headerList}
\usage{
blankCorrection_variable_headerList(
  grouptable,
  intensityFactor = 10,
  head = headerList
)
}
\arguments{
\item{grouptable}{Alignmenttabelle der grouping_BfG2 Funktion}

\item{intensityFactor}{Faktor für die Intensitätsdifferenz zwischen 2 Peaks (Vgl. grouping_BfG2)}

\item{head}{headerlist aus der Non-target App}
}
\value{
Blankkorrigierte Alignmenttabelle
}
\description{
Blankabzug nach Vorbild der orginalen Non-target App Funktion mit Berücksichtigung der Gruppenzugehörigkeit
 
Ist ein Peak im Blank wird IN JEDER Probe je nach seiner Intensität in der Gruppe entschieden:
  Int_max der Gruppe: -> die ganze Gruppe wird in dieser Probe entfernt
  nicht Int_max der Gruppe: -> nur dieser Peak wird in der Probe entfernt
}
