% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library-screening.R
\name{calc_ndp}
\alias{calc_ndp}
\title{Calculate Normalized Dot-Product}
\usage{
calc_ndp(d_spec, db_spec, ndp_m = 2, ndp_n = 1, mztolu_ms2 = 0.015)
}
\arguments{
\item{ndp_m}{Peak intensity weighting factor}

\item{ndp_n}{m/z weighting factor}

\item{data_spec}{Spectrum from data}

\item{stan_spec}{Spectrum from database}

\item{mztolu}{m/z tolerance for binning peaks (u)}
}
\value{
A numeric from 0 to 1000 with 1000 a perfect match
}
\description{
Calculate Normalized Dot-Product
}
\details{
See MetCirc::NDP for more information
}
