% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/componentization.R
\name{componentization_BfG}
\alias{componentization_BfG}
\title{componentization_BfG}
\usage{
componentization_BfG(
  Liste,
  daten,
  ppm = 10,
  Grenzwert_RT = 10,
  Grenzwert_FWHM_left = 10,
  Grenzwert_FWHM_right = 10,
  Summe_all = 10,
  adjust_tolerance = TRUE
)
}
\arguments{
\item{Liste}{(peaklist)}

\item{daten}{(datenList)}

\item{ppm}{(mass deviation in ppm)}

\item{Grenzwert_RT}{(Threshold value for the retention time)}

\item{Grenzwert_FWHM_left}{(Limit value for the retention time of the left FWHM)}

\item{Grenzwert_FWHM_right}{(Limit value for the retention time of the right FWHM)}

\item{Summe_all}{(Threshold value for the sum of the three retention times (apex, FWHM left, FWHM right)}

\item{adjust_tolerance}{(adjust thresholds with 13C isotopologue)}
}
\value{
peaklist with added cols groupleader and group for each component
}
\description{
Componentization/ Grouping of adducts, isotopologues, in-source fragments, etc.
Either by:
 fixed thresholds for RT, RT_FWHM_left and RT_FWHM_right or
 automatically adjusted thresholds based on the 13C isotopologue.
}
