% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{correct_formula}
\alias{correct_formula}
\title{Adjust a formula to an adduct}
\usage{
correct_formula(formula, adduct)
}
\arguments{
\item{adduct}{}
}
\value{

}
\description{
Adjust a formula to an adduct
}
