% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak-picking-alignment.R
\name{correlates_with_r}
\alias{correlates_with_r}
\title{Find Pearson's correlation between intensity trends}
\usage{
correlates_with_r(aligned_intensities, zeile, koeffizient)
}
\arguments{
\item{aligned_intensities}{Matrix of intensities}

\item{zeile}{Reference row of matrix to which all other are compared to}

\item{koeffizient}{Pearson correlation coefficient (minimum)}
}
\value{
Boolean. True for those rows of the matrix which have a correlation 
coefficient higher than the threshold.
}
\description{
Wrapper for the cpp function correlates_with. Takes a matrix of intensities
(each row representing one feature) and a row, returns which rows have
correlation coefficients within threshold
}
