% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Label_db.R
\name{create_DB_json}
\alias{create_DB_json}
\title{create_DB_json}
\usage{
create_DB_json(
  samplelist_i = sampleList,
  json_path = "~/Data/Toni/Lable_Datenbank/Json_Files",
  db_path = "~/sqlite_local",
  grouped_i = grouped,
  peaklist_i = peaklist,
  annotationTable_i = annotationTable,
  datenlist_i = datenList
)
}
\arguments{
\item{json_path}{path to json-files on Z}

\item{db_path}{path to both DBs on Server (Spectra_DB = db_v7 and Label_DB = db_v5)}

\item{sampleList}{from NT-App/ environment}

\item{grouped}{alignmentTable from NT-App/ environment}

\item{peaklist}{from NT-App/ environment}

\item{annotationTable}{from NT-App/ environment}
}
\value{
save Json for Label_DB from NT-App Data
}
\description{
create_DB_json
}
