% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{create_log_file}
\alias{create_log_file}
\title{create_log_file}
\usage{
create_log_file()
}
\arguments{
\item{no}{parameters needed}
}
\value{
log_file with used settings for peakpicking, alignment, blankcorrectio,normalization, filtering, annotation
}
\description{
create_log_file
}
