% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library-screening.R
\name{dbGetSpectrum}
\alias{dbGetSpectrum}
\title{Extraction of a MS2-Spectrum for a given ExperimentID from spectral library}
\usage{
dbGetSpectrum(db, Exp.ID)
}
\arguments{
\item{db}{database connection object using dplyr::src_sqlite}

\item{Exp.ID}{integer of experiment ID (only 1)}
}
\value{
MS2-Spectrum as data.frame (mz, int)
}
\description{
Extraction of a MS2-Spectrum for a given ExperimentID from spectral library
}
