% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{Einleiter}
\alias{Einleiter}
\title{Einleiter}
\usage{
Einleiter(
  peaklist_e = new_peaklist,
  Proben_Auswahl_Vector = c(1:length(new_peaklist)),
  ppm = 10,
  DeltaRT = 10,
  Int_Lim = 10,
  intensityFactor = 10
)
}
\arguments{
\item{peaklist_e}{Peaklist aus der Non-target App oder auch aus Gruppierung_SG}

\item{Proben_Auswahl_Vector}{Vektor welche Proben untersucht werden sollen}

\item{ppm}{erlaubte Massendifferenz zwischen 2 Peaks}

\item{DeltaRT}{erlaubte RT-differenz zwischen 2 Peaks}

\item{Int_Lim}{Intensitätsthreshold ab wann Peaks untersucht werden sollen
Peaks unterhalb des Thresholds kommen nicht in die Ergebnisliste}

\item{intensityFactor}{Faktor für die Intensitätsdifferenz zwischen 2 Peaks (Vgl. grouping_BfG2)}
}
\value{
Peaklist mit der Information welche Peaks zwischen 2 Proben unterschiedlich sind, Einleiter halt
}
\description{
Funktion zur Ermittlung "neuer" Peaks einer Probe im Vergleich zur vorherigen Probe
}
\details{
Hinweis: wenn man das Ergebnis noch alignen will sollte man die Werte für ppm und DeltaRT in beiden Funktionen angleichen,
          da sonst u.U. Werte alignt werden, welche laut Einleiter eigentlich entfernt sein sollten!
2. Hinweis: Es werden NUR aufeinander folgende Proben verglichen!
            Wenn ein Peak in der vorherigen Probe (im Code: i-1) nicht vorkommt, aber in 2 Proben davor (sinngemäß i-2) schon, 
            wird er als "neu" in die Ergebnisliste aufgenommen
            Alternative -> Funktion Einleiter_alles_vorher
}
