% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{Einleiter_alles_vorher}
\alias{Einleiter_alles_vorher}
\title{Einleiter_alles_vorher}
\usage{
Einleiter_alles_vorher(
  peaklist_e = new_peaklist,
  Proben_Auswahl_Vector = c(1:length(new_peaklist)),
  ppm = 10,
  DeltaRT = 10,
  Int_Lim = 10,
  intensityFactor = 10
)
}
\arguments{
\item{peaklist_e}{Peaklist aus der Non-target App oder auch aus Gruppierung_SG}

\item{Proben_Auswahl_Vector}{Vektor welche Proben untersucht werden sollen}

\item{ppm}{erlaubte Massendifferenz zwischen 2 Peaks}

\item{DeltaRT}{erlaubte RT-differenz zwischen 2 Peaks}

\item{Int_Lim}{Intensitätsthreshold ab wann Peaks untersucht werden sollen
Peaks unterhalb des Thresholds kommen nicht in die Ergebnisliste}

\item{intensityFactor}{Faktor für die Intensitätsdifferenz zwischen 2 Peaks (Vgl. grouping_BfG2)}
}
\value{
Peaklist mit der Information welche Peaks zwischen einer Proben und allen vorherigen unterschiedlich sind
}
\description{
Funktion zur Ermittlung "neuer" Peaks einer Probe im Vergleich zu ALLEN vorherigen Proben in der Peaklist
Hinweis: wenn man das Ergebnis noch alignen will sollte man die Werte für ppm und DeltaRT in beiden Funktionen angleichen,
         da sonst u.U. Werte alignt werden, welche laut Einleiter eigentlich entfernt sein sollten!
         
kann bei vielen Proben zu leeren Listen führen, wenn alle Peaks einer Probe bereits in vorherigen Proben gemessen wurden
 -> führt auf die Weise zum Absturz der Alignmentfunktion
}
