% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak-picking-alignment.R
\name{FindPeaks_BfG}
\alias{FindPeaks_BfG}
\title{Peak finding algorithm}
\usage{
FindPeaks_BfG(
  daten,
  mz_min,
  mz_max,
  mz_step,
  rt_min,
  rt_max,
  sn,
  int_threshold,
  peak_NoiseScans,
  precursormzTol,
  peakwidth_min,
  peakwidth_max,
  maxPeaksPerSignal
)
}
\arguments{
\item{daten}{from xcms}

\item{mz_min}{}

\item{mz_max}{}

\item{mz_step}{}

\item{rt_min}{}

\item{rt_max}{}

\item{sn}{}

\item{int_threshold}{}

\item{peak_NoiseScans}{}

\item{precursormzTol}{}

\item{peakwidth_min}{}

\item{peakwidth_max}{}

\item{maxPeaksPerSignal}{}
}
\value{
data.frame of the peak inventory list
}
\description{
Function will pick chromatographic peaks in the raw data file by binning and 
analyzing the extracted ion chromatograms. The peak finding is done by
taking the derivative of the chromatogram and looking for points where it 
crosses 0. More information can be found in Dietrich, C., Wick, A., & 
Ternes, T. A. (2021). Open source feature detection for non‐target LC‐MS 
analytics. Rapid Communications in Mass Spectrometry, e9206. doi:10.1002/rcm.9206
}
