% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakPicking.R
\name{FindPeaks_BfG_scripting}
\alias{FindPeaks_BfG_scripting}
\title{Peak picking function simplified for scripting}
\usage{
FindPeaks_BfG_scripting(
  daten,
  mz_min,
  mz_max,
  mz_step,
  rt_min,
  rt_max,
  sn,
  int_threshold,
  peak_NoiseScans,
  precursormzTol,
  peakwidth_min,
  peakwidth_max,
  maxPeaksPerSignal
)
}
\arguments{
\item{daten}{}

\item{mz_min}{}

\item{mz_max}{}

\item{mz_step}{}

\item{rt_min}{}

\item{rt_max}{}

\item{sn}{}

\item{int_threshold}{}

\item{peak_NoiseScans}{}

\item{precursormzTol}{}

\item{peakwidth_min}{}

\item{peakwidth_max}{}

\item{maxPeaksPerSignal}{}
}
\value{

}
\description{
The original peak picking function can only be used together with the app.
This function can be used in a scripting environment. Results are the same.
This is not parallelized, since it assumed to calling function is already parallelized.
}
