% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_extraction.R
\name{FindPeaks_SingleXIC2}
\alias{FindPeaks_SingleXIC2}
\title{Christian's FindPeaks algorithm}
\usage{
FindPeaks_SingleXIC2(
  i,
  daten,
  mz_step,
  rt_min_scan,
  rt_max_scan,
  sn,
  int_threshold,
  peak_NoiseScans,
  precursormzTol
)
}
\arguments{
\item{precursormzTol}{}
}
\description{
Christian's FindPeaks algorithm
}
