% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{forIdentExport}
\alias{forIdentExport}
\title{Export Function to For-Ident.}
\usage{
forIdentExport(
  align_matrix,
  sampleList,
  rawDataList,
  intThreshRel = 0.01,
  report_nm
)
}
\arguments{
\item{align_matrix}{Alignment \code{matrix} from Christian's App}

\item{sampleList}{sampleList \code{matrix} from App}

\item{rawDataList}{datenList \code{list} from App}

\item{intThreshRel}{Intensity threshold for including peaks in MS2 spectra}

\item{report_nm}{Name of output file}
}
\value{
A txt file with each found mass, retention time and MS2 spectrum (if available).
}
\description{
For batch processing in For-Ident, txt file is created for import on For-Ident platform. MS2 
spectra are included for Met-frag or Massbank search. The spectrum for each row is taken from the
sample with the highest intensity in this row. The function takes approx. 1 min to process 100 
rows, so limit the size of the alignment matrix passed to the app. If no MS2 is available, row is
skipped.
}
