% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suspect_screening.R
\name{get_ids}
\alias{get_ids}
\title{Get Experiment IDs From DB By Mz and Rt}
\usage{
get_ids(
  db,
  mz,
  rt,
  mztolu = 0.005,
  rttol = 1,
  ExpGroupNames = c("BfG"),
  compGroupNames = c("BfG"),
  polarity = "pos",
  ionisation = "ESI",
  CE = 35,
  CES = 15,
  ce_unit = "V",
  col_type = "Q",
  instrument = "LC-ESI-QTOF TripleTOF 5600 SCIEX",
  chrom_method = "dx.doi.org/10.1016/j.chroma.2015.11.014"
)
}
\arguments{
\item{chrom_method}{}
}
\value{

}
\description{
Can't include ExpGroupNames due to stack overflow, so this parameter doesn't
do anything yet
}
