% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suspect_screening.R
\name{get_MS2spectrum}
\alias{get_MS2spectrum}
\title{Extracting Spetra from mzML, mzXML data}
\usage{
get_MS2spectrum(data.mzR, mass, rt, p_mztolu = 0.5, p_rttol = 0.2)
}
\arguments{
\item{data.mzR}{object resulting from mzR::openMSfile}

\item{mass}{m/z of MS1 peak}

\item{rt}{retention time of MS1 peak in min}

\item{p_mztolu}{mz tol in u}

\item{p_rttol}{rt tol in min}
}
\value{
spectrum as data.frame (headers: mz, int), if no MS2 is found, returns NULL
}
\description{
Extracting Spetra from mzML, mzXML data
}
