% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{getTrends}
\alias{getTrends}
\title{get Trends}
\usage{
getTrends(
  grouptable,
  Int_Columns = grep("Int_", colnames(grouptable)),
  sn = 5,
  Int_Threshold = 20
)
}
\arguments{
\item{grouptable}{von Alignment oder Blankabzug}

\item{Int_Columns}{Int Spalten der grouptable}

\item{sn}{Signal/ Noise minimum}

\item{Int_Threshold}{min Int der Maxima}
}
\value{
In welchen Proben haben Peaks ihre Intensitätsmaxima und in welchen
        dazugehörigen Proben beginnt der Anstieg der Intensität
}
\description{
Trenderkennung mit Angabe der Start und Endpunkte (Proben) des Intensitätsanstiegs
}
