% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{grouping_BfG_ohne_datenlist}
\alias{grouping_BfG_ohne_datenlist}
\title{grouping_BfG_ohne_datenlist}
\usage{
grouping_BfG_ohne_datenlist(peaklists, ppm = 10, DeltaRT = 10, ms2ppm = 10)
}
\arguments{
\item{peaklists}{Peaklisten aus der Funktion Gruppierung_SG}

\item{ppm}{erlaubte Massendifferenz zwischen 2 Peaks}

\item{DeltaRT}{erlaubte RT-differenz zwischen 2 Peaks}

\item{ms2ppm}{(DEAKTIVIERT) erlaubte Massendifferenz im MS2-Spektrum zwischen 2 Peaks}
}
\value{
Alignmenttabelle mit zusätzlicher Gruppeninformation
}
\description{
Funktion für das Alignment der Ergebnispeaklist der Gruppierungsfunktion Gruppierung_SG
Die Spalte Gruppe wird für jede Probe in die Alignmenttabelle integriert
der MS2 Vergleich ist deaktiviert, da die Peaklist in der Funktion Replikate gekürzt wirde und ein Verweis in
die Datenlist der Non-target App fehlschlägt
}
