% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{Gruppierung_SG}
\alias{Gruppierung_SG}
\title{Gruppierung_Summe_Grenzwert}
\usage{
Gruppierung_SG(
  peaklist_df,
  mit_RT = TRUE,
  Grenzwert_RT = 1.5,
  Grenzwert_FWHM_left = 0.5,
  Grenzwert_FWHM_right = 1,
  Summe_all = 2.5,
  Summe_FWHM = 1.5
)
}
\arguments{
\item{peaklist_df}{Peaklist aus der Non-target App, oder aus der Funktion Replikate, egal, hauptsache Peaklist}

\item{mit_RT}{soll die RT für die Gruppierung mit benutzt weden (TRUE), oder nur die FWHM_left und FWHM_right (FALSE)}

\item{Grenzwert_RT}{erlaubte RT-differenz zwischen 2 Peaks}

\item{Grenzwert_FWHM_left}{erlaubte FWHM_left-differenz zwischen 2 Peaks}

\item{Grenzwert_FWHM_right}{erlaubte FWHM_right-differenz zwischen 2 Peaks}

\item{Summe_all}{(bei mit_RT=TRUE) erlaubte maximale Summe aus RT-diff, FWHM_left-diff und FWHM_right-diff zwischen 2 Peaks}

\item{Summe_FWHM}{(bei mit_RT=FALSE) erlaubte maximale Summe aus FWHM_left-diff und FWHM_right-diff zwischen 2 Peaks}
}
\value{
Peaklist mit der Information der Gruppenzuordnung
}
\description{
ordnet die Peaks in Gruppen (beginnend bei den intensivsten)
die Gruppen sollten neben dem [M+H]+ Peak die Addukte, Isotope und Fragmente enthalten
Die Spalte "Gruppe" wird an die Peaklist angehangen
}
\details{
Für das Alignment wird die Funktion grouping_BfG2 benötigt
}
