% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Label_db.R
\name{insert_json_data}
\alias{insert_json_data}
\title{insert_json_data}
\usage{
insert_json_data(
  json_path = "~/Data/Toni/Lable_Datenbank/Json_Files",
  db_path = "~/sqlite_local"
)
}
\arguments{
\item{json_path}{path to json-files on Z}

\item{db_path}{path to both DBs on Server (Spectra_DB = db_v7 and Label_DB = db_v5)}
}
\value{
insert json-files to Label_DB (db_v5)
}
\description{
insert_json_data
}
