% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment-table-filters.R
\name{keep_reps_regex}
\alias{keep_reps_regex}
\title{Replicate feature based on regular expression}
\usage{
keep_reps_regex(alignment, samples, sampleList, regexp, least)
}
\arguments{
\item{alignment}{alignment table (matrix, grouped)}

\item{samples}{Sample IDs which should be considered (e.g. only samples and no blanks)}

\item{sampleList}{Sample list dataframe with columns ID, File etc.}

\item{regexp}{string length 1 regular expression by which to group the samples into replicates
this is done removing the matching string with stringr::str_replace(filename, regexp, "\\1"). The 
part of the regexp in brackets is kept (\\1) and so should be the same in all replicates.}

\item{least}{How many times does the feature need to appear. Must be less than the number of
replicates in each group. If it is higher, will be changed to the number of replicates in the group with a message.}
}
\value{
alignment table with rows removed where the number of aligned features 
in each replicate set is less than 'least'.
}
\description{
use a regular expression on the filename to gather the groups of samples in the batch into
sets. Then apply the keepReps function on each of these sets in turn. This will be quite slow
because we are looping through the sets and each set loops through each row. But it is the
easiest way for now.
}
