% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentTableFilters.R
\name{keepConsecutive}
\alias{keepConsecutive}
\title{Keep only rows found in at least X consecutive samples}
\usage{
keepConsecutive(alignment, samples, consecutive)
}
\arguments{
\item{alignment}{Alignment Table}

\item{samples}{samples to consider (everything else will be ignored)}

\item{consecutive}{number of consecutive samples a peak should be present in}
}
\value{

}
\description{
Keep only rows found in at least X consecutive samples
}
