% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentTableFilters.R
\name{keepReps}
\alias{keepReps}
\title{Only retain peaks found in all replicates}
\usage{
keepReps(alignment, samples, reps, least)
}
\arguments{
\item{alignment}{alignment table}

\item{samples}{numeric vector sample IDs to merge, length must be divisable by reps}

\item{reps}{numer of replicates}

\item{least}{Peak found in at least this many replicates, must be <= reps}
}
\value{
An alignment table with the same number of columns but a reduced number of rows.
}
\description{
This function will check, for a given number of replicate injections, that a feature was found
in all replicates. Features will be set to zero if they are not found in replicates.
Aligned features not found in any samples (everything zero) will be removed.
}
