% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{loadReport}
\alias{loadReport}
\title{Load a previously saved Report}
\usage{
loadReport(dialog = TRUE, path = NULL)
}
\arguments{
\item{dialog}{logical should a dialog window open to select the file (only works in rstudio)}

\item{path}{if dialog is false a path to the file. Must be given if rstudioapi is not available}
}
\value{
an `ntsworkflow::Report` object
}
\description{
A dialog box is used or the path can be given as an argument.
}
\details{
rstudioapi is used for the dialog. readRDS can be used directly to open saved Report objects as well.
}
\seealso{
readRDS
}
