% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakPicking.R
\name{new_peaklist}
\alias{new_peaklist}
\title{Correct peaklists for parallel computing}
\usage{
new_peaklist(peaklist, datenList)
}
\arguments{
\item{peaklists}{}
}
\value{
New peaklist list in the same form as before but with additional columns
sample_id and peak_id_all. If these were already present, they are deleted and re-written
}
\description{
Correct peaklists for parallel computing
}
