% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library-screening.R
\name{normalizeMs2}
\alias{normalizeMs2}
\title{Prepare Spectrum For Comparison}
\usage{
normalizeMs2(x)
}
\arguments{
\item{x}{spectrum as data.frame columns mz, int}
}
\value{
spectrum with intensities to max intensity and precursor mass removed
}
\description{
Prepare Spectrum For Comparison
}
