% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntsworkflow-package.R
\docType{package}
\name{ntsworkflow-package}
\alias{ntsworkflow-package}
\title{ntsworkflow: A Non-Target Screening Data Processing Tool. Federal Institute of Hydrology, Koblenz, Germany}
\description{
ntsworkflow is a collection of functions for non-target-screening measurment data processing, including peak-picking, alignment and annotation.
}
\author{
\strong{Maintainer}: Kevin S. Jewell \email{Jewell@bafg.de} (Federal Institute of Hydrology (BfG) Koblenz, Germany)

Authors:
\itemize{
  \item Christian Dietrich \email{Christian.Dietrich@bafg.de} (Federal Institute of Hydrology (BfG) Koblenz, Germany)
  \item Toni Köppe \email{Koeppe@bafg.de} (Federal Institute of Hydrology (BfG) Koblenz, Germany)
  \item Franziska Thron \email{Thron@bafg.de} (Federal Institute of Hydrology (BfG) Koblenz, Germany)
  \item Arne Wick \email{Wick@bafg.de} (Federal Institute of Hydrology (BfG) Koblenz, Germany)
}

}
\keyword{internal}
