% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentTableFilters.R
\name{onlyGroupLeaders}
\alias{onlyGroupLeaders}
\title{Keep only features which are the group leaders}
\usage{
onlyGroupLeaders(alignment, pll, samples)
}
\arguments{
\item{alignment}{alignment table}

\item{pll}{peaklistlist}

\item{samples}{sample IDs of the samples which will be looked at, all other samples will be ignored}
}
\value{
an alignment table with reduced number of rows
}
\description{
This will go through the alignment table and check every feature if it is a group leader. Features
which are marked as group leader in ANY sample will kept. In other words, if the componentization is inconsistent
and a feature is sometimes marked as groupleader and sometimes not, the feature will be kept.
}
