% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak-picking.R
\name{optimumMzStep}
\alias{optimumMzStep}
\title{Get a recommendation for the m/z-step parameter (binning width)}
\usage{
optimumMzStep(daten, probs)
}
\arguments{
\item{daten}{raw data as list of xcmsRAW objects}

\item{probs}{quantile to use (see quantile function)}
}
\value{
Recommended m/z width (Da)
}
\description{
Calculates the optimum mz step based on mass differences. Takes a sample
of 100 spectra to make the calculation
}
