% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak-picking-alignment.R
\name{peakpicking_BfG_cpp}
\alias{peakpicking_BfG_cpp}
\title{Peakpeaking algorithm using c++}
\usage{
peakpicking_BfG_cpp(
  i,
  rawData,
  mz_step,
  rt_min_scan,
  rt_max_scan,
  sn,
  int_threshold,
  NoiseScans,
  peakwidth_min,
  peakwidth_max,
  precursormzTol,
  maxPeaksPerSignal
)
}
\arguments{
\item{i}{lower mz for extracted ion chromatogram}

\item{rawData}{}

\item{mz_step}{mz width for extracted ion chromatogram}

\item{rt_min_scan}{}

\item{rt_max_scan}{}

\item{sn}{}

\item{int_threshold}{}

\item{NoiseScans}{}

\item{peakwidth_min}{}

\item{peakwidth_max}{}

\item{precursormzTol}{}

\item{maxPeaksPerSignal}{}
}
\value{
matrix of peaks detected in the extracted ion chromatogram, each peak
is one row with the columns representing various parameters for the peak.
}
\description{
This function wraps the c++ peakpicking algorithm. The function picks peaks
in a specific mz range (defined by i and mz_step). Used by the function
FindPeaks_BfG.
}
