% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{Replikate}
\alias{Replikate}
\title{Replikate}
\usage{
Replikate(peaklist_r = peaklist, ppm = 5, DeltaRT = 3, Rep)
}
\arguments{
\item{peaklist_r}{aus der Non-target App}

\item{ppm}{erlaubte Massendifferenz zwischen 2 Peaks}

\item{DeltaRT}{erlaubt RT-differenz zwischen 2 Peaks}

\item{Rep}{Anzahl der zu einer Probe gehörenden Replikate (z.B. Triplikate, Rep=3)}
}
\value{
Peaklist mit vereinigten Replikaten und Blanks
}
\description{
Diese Funktion vereinigt Replikate in der Messung unter der Vorgabe, dass ein Peak in allen Replikaten vorkommen muss
 Dabei sollen Blanks nicht wereinigt werden und müssen deswegen mit "Anzahl_Blank" angegeben und am Ende der
 input-peaklist stehen
 
 Nur für Rep=2 bis Rep=4 programmiert!
}
\examples{
Messung von 5 Proben als Triplikat mit 3 Blanks
          Rep = 3, Anzahl_Proben = 15, Anzahl_Blank = 3
}
