% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPeakPicking.R
\name{runPeakPicking}
\alias{runPeakPicking}
\title{Start the shiny peak picking app from the R console}
\usage{
runPeakPicking()
}
\description{
The current working directory will become the working directory for the app.
The app will display as a browser pop-up.
}
