% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakPicking.R
\name{summarize_groups}
\alias{summarize_groups}
\title{Summarise groups into one column (very experimental feature)}
\usage{
summarize_groups(alig)
}
\description{
Spalte Gruppe (Zusammenfassung der Gruppen über die gesamte Alignmenttabelle) wird hier ausgefüllt
}
