% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{trueInaRow}
\alias{trueInaRow}
\title{Search for trues in a row}
\usage{
trueInaRow(set, in_a_row = 2)
}
\arguments{
\item{set}{logical}

\item{in_a_row}{}
}
\value{
Logical of length 1
}
\description{
Search for trues in a row
}
