% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_class.R
\docType{class}
\name{UfidReport-class}
\alias{UfidReport-class}
\alias{UfidReport}
\title{A reference class for extracting ufids from raw data for import into elastic search}
\description{
This class extends Report class for use with the ufid database
The peak list has an additional column for the ufid
}
\section{Methods}{

\describe{
\item{\code{ufid_search_one(data_raw, db_con, select_ufids = NULL)}}{This function takes an xcmsRaw file and looks for peaks found in the ufid db
It needs a DBI connection to the ufid database and returns a peakList of 
the form returned by process_all}
}}

