% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{wind}
\alias{wind}
\title{Simple function to create window tolerance}
\usage{
wind(value, tol)
}
\arguments{
\item{value}{}

\item{tol}{}
}
\value{
two numbers
}
\description{
Internal package use only
}
