% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak-picking-alignment.R
\name{alignment_BfG_cpp}
\alias{alignment_BfG_cpp}
\title{Alignment of peaks from difference samples}
\usage{
alignment_BfG_cpp(peaklists, mz_dev, DeltaRT, mz_dev_unit)
}
\arguments{
\item{peaklists}{list of peaklists}

\item{mz_dev}{m/z tolerance}

\item{DeltaRT}{rt tolerance in s}

\item{mz_dev_unit}{Units for m/z tolerance (ppm or mDa)}
}
\value{
Matrix of aligned peaks, one aligned feature per row
}
\description{
This a wrapper for the C++ function alignmentBfGC. It takes a list of peaklists
and produces an alignment table.
}
