% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak-picking-alignment.R
\name{alignment_BfG_cpp_par}
\alias{alignment_BfG_cpp_par}
\title{Align peaks with parallel processing}
\usage{
alignment_BfG_cpp_par(
  peaklists,
  ppm_dev,
  DeltaRT,
  mz_dev_unit,
  mDa_split = 100,
  numSplits = 16
)
}
\arguments{
\item{peaklists}{list of peaklists}

\item{ppm_dev}{mz tolerance}

\item{DeltaRT}{rt tolerance in s}

\item{mz_dev_unit}{units for mz tolerance (ppm or mDa)}

\item{mDa_split}{mz gap between chunks}

\item{numSplits}{Number parallel chunks to create}
}
\value{
matrix of aligned peaks, one aligned feature per row
}
\description{
To align very large datasets it is necessary to split the alignment process
into parallel jobs. This function slips the peak lists and runs the
alignment_BfG_cpp on each chunk in parallel. The results are combined. Uses
forking therefore not available on Windows.
}
