% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suspect_screening.R
\name{annotate_grouped}
\alias{annotate_grouped}
\title{Annotate alignment table from non-target app}
\usage{
annotate_grouped(
  sampleListLocal,
  peakListList,
  alignmentTable,
  db_path,
  threshold_score = 500,
  mztolu = 0.005,
  rttol = 1,
  polarity = "pos",
  CE = c(30, 40),
  CES = c(0, 15),
  instrument = c("LC-ESI-QTOF TripleTOF 5600 SCIEX",
    "LC-ESI-QTOF TripleTOF 6600 SCIEX", "LC-ESI-Orbitrap QExactive"),
  chrom_method = "dx.doi.org/10.1016/j.chroma.2015.11.014",
  ndp_m = 2,
  ndp_n = 1,
  mztolu_ms2 = 0.015,
  rtoffset = 0,
  intCutData = 0
)
}
\arguments{
\item{peakListList}{}

\item{alignmentTable}{}

\item{db_path}{}

\item{threshold_score}{}

\item{mztolu}{}

\item{rttol}{}

\item{polarity}{}

\item{CE}{}

\item{CES}{}

\item{instrument}{}

\item{chrom_method}{}

\item{ndp_m}{}

\item{ndp_n}{}

\item{mztolu_ms2}{}

\item{rtoffset}{Note: Offset the database to make it match your values}

\item{intCutData}{}
}
\value{

}
\description{
Function takes all the variables needed from non-target app and settings for 
searching the database. It returns a table with the annotations for rows of 
the alignment table.
}
\details{
datenList is taken from parent frame

If data files are not in memory, they will be temporary loaded and
}
