% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/componentization.R
\name{annotate_grouped_components}
\alias{annotate_grouped_components}
\title{Annotate alignment table with adduct and isotopologue information}
\usage{
annotate_grouped_components(
  sampleListLocal,
  peakListList,
  alignmentTable,
  numcores
)
}
\arguments{
\item{sampleListLocal}{(`data.frame` of processed samples)}

\item{peakListList}{(from peakpicking)}

\item{alignmentTable}{(from alignment)}

\item{numcores}{(number of CPU cores to be used)}
}
\value{
alignmentTable with annotated isotopologues based on componentization
}
\description{
This function is for the annotation section of the app. If no db
  is chosen, it will annotate the alignmentTable with peaklist component info
  (Cl and Br and some adducts)
}
