% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentTableFilters.R
\name{blankCorrection}
\alias{blankCorrection}
\title{Remove features also found in the blanks}
\usage{
blankCorrection(
  alignment,
  samplels,
  intensityFactor = 10,
  deleteGrouped = TRUE
)
}
\value{
an alignment table with rows removed
}
\description{
Remove features also found in the blanks
}
