% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{correct_formula}
\alias{correct_formula}
\title{Combine a formula with the adduct}
\usage{
correct_formula(formula, adduct)
}
\arguments{
\item{formula}{Chemical formula as a string}

\item{adduct}{adduct as a string, must be one of [M+H]+, [M+Na]+, [M]+,
[M-H]-, [M+NH4]+, [M+H2CO2-H]-, [M-H2O+H]+}
}
\value{
List with two elements: formula with the adduct added to it and the charge
}
\description{
Combine a formula with the adduct
}
