% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mergeReport}
\alias{mergeReport}
\title{Merge two Report objects}
\usage{
mergeReport(report1, report2)
}
\arguments{
\item{report1}{A Report object}

\item{report2}{A second Report object}
}
\value{
an ntsworkflow::Report object
}
\description{
The function will combine two report objects into one. The peak ids of the first
object are followed by the second.
}
