% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suspect_screening.R
\name{ms2_search}
\alias{ms2_search}
\title{Suspect search without peak searching, looking only at MS2 spectra}
\usage{
ms2_search(
  data_path,
  db_path,
  rttolm = 1,
  mztolu = 0.5,
  mztolu_fine = 0.005,
  chromatography = "dx.doi.org/10.1016/j.chroma.2015.11.014",
  pol = "pos",
  CE_s = 30:40,
  CES_s = 0:15,
  instr = "LC-ESI-QTOF TripleTOF 5600 SCIEX",
  ceunit = c("V", "eV"),
  comparison = "dot_product",
  threshold = 400,
  rt_res = 1.5,
  rtoffset = 0,
  ndp_m = 2,
  ndp_n = 1,
  mztolu_ms2 = 0.015,
  compounds = NULL
)
}
\arguments{
\item{data_path}{Path to rawfiles, can be a vector of file locations or a list of xcmsRaw objects}

\item{db_path}{Path to spectral library}

\item{rttolm}{Retention time tolerance in minutes}

\item{mztolu}{m/z tolerance in Da (MS2 precursor mass)}

\item{mztolu_fine}{m/z tolerance in Da (second stage)}

\item{chromatography}{Chromatography name to choose the correct retention times from the library}

\item{pol}{Polarity used}

\item{CE_s}{Collision energy}

\item{CES_s}{Collision energy spread}

\item{instr}{Instruments allowed, must match the instruments listed in the library}

\item{ceunit}{Collision energy units}

\item{comparison}{How the spectra should be compared, at the moment only "dot_product" is allowed}

\item{threshold}{Threshold score for spectral comparison (for dot_product 1000 is max)}

\item{rt_res}{resolution of two peaks in chromatography}

\item{rtoffset}{Retention time offset (will be added to library RTs)}

\item{ndp_m}{Peak intensity weighting factor for dot-product}

\item{ndp_n}{m/z weighting factor for dot-product}

\item{mztolu_ms2}{m/z window for dot-product}

\item{compounds}{character vector of compounds which should be processed}
}
\description{
Suspect search without peak searching, looking only at MS2 spectra
}
