% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak-picking-alignment.R
\name{new_peaklist}
\alias{new_peaklist}
\title{Add additional peak and sample ID information to peaklists}
\usage{
new_peaklist(peaklist, datenList)
}
\arguments{
\item{peaklist}{}

\item{datenList}{}
}
\value{
New peaklist list in the same form as before but with additional columns
sample_id and peak_id_all. If these were already present, they are deleted and re-written
}
\description{
This needs to be run before alignment
}
