% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakPicking.R
\name{optimumMzStep}
\alias{optimumMzStep}
\title{Calculates the optimum mz step based on mass differences. Takes a sample
of 100 spectra to make the calculation}
\usage{
optimumMzStep(daten, probs)
}
\description{
Calculates the optimum mz step based on mass differences. Takes a sample
of 100 spectra to make the calculation
}
