% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentTableFilters.R
\name{removeRare}
\alias{removeRare}
\title{Remove features found in less than x samples}
\usage{
removeRare(alignment, minimum)
}
\arguments{
\item{alignment}{alignment table}

\item{minimum}{number of samples in which a feature is found must be less than
total number of samples}
}
\value{
alignment table with rows removed
}
\description{
Remove features found in less than x samples
}
