% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak-picking-alignment.R
\name{summarize_groups}
\alias{summarize_groups}
\title{Summarize componentization groups into one column}
\usage{
summarize_groups(alig)
}
\arguments{
\item{alig}{Alignment table}
}
\value{
Matrix. Alignment table with additional column Gruppe
}
\description{
Will take the componentization information from each sample in the
alignment table and attempt to summerize this into one column named "Gruppe".
}
