% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{wind}
\alias{wind}
\title{Simple function to create window tolerance}
\usage{
wind(value, tol)
}
\arguments{
\item{value}{Central number, e.g. m/z}

\item{tol}{Tolerance}
}
\value{
numeric length 2 with start and end of window
}
\description{
Simple function to create window tolerance
}
